add: e2e test and remove unneeded tests
