--- e2e-tests/src/lib.rs
@@ -1,12 +1,0 @@
-//! E2E Tests Library
-//!
-//! This crate contains end-to-end tests for the SNOS workspace.
-//! It's a separate crate to properly handle dependencies and test configuration.
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
